+++ Assets/QA/Input_Test/Scripts/Input/TouchISX.cs
@@ -17,7 +17,7 @@ public class TouchISX : MonoBehaviour
     // Use this for initialization
     void Start()
     {
-        m_touchAction = new InputAction(name: "TouchAction", binding: "<touchscreen>/<touch>");
+        m_touchAction = new InputAction(name: "TouchAction", binding: "<touchscreen>/touch*"); // Not using <touch> in order to not get primaryTouch, too.
         m_touchAction.performed += callbackContext => TouchInput(callbackContext.control as TouchControl);
         m_touchAction.canceled += callbackContext => TouchInput(callbackContext.control as TouchControl);
         m_touchAction.Enable();