+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/StandaloneInputModuleEditor.cs
@@ -8,7 +8,7 @@ namespace UnityEngine.InputSystem.UI.Editor
     // The only purpose of the Input System suppying a custom editor for the UI StandaloneInputModule is to guide users to using
     // the Input System's InputSystemUIInputModule instead.
     [CustomEditor(typeof(StandaloneInputModule))]
-    internal class StandaloneInputModuleModuleEditor : UnityEditor.Editor
+    internal class StandaloneInputModuleEditor : UnityEditor.Editor
     {
         SerializedProperty enableNativePlatformBackendsForNewInputSystem;
         SerializedProperty disableOldInputManagerSupport;